# [aggregator] Reduce timer stream allocs